+++ src/main/java/org/opensearch/neuralsearch/query/visitor/NeuralSearchQueryVisitor.java
@@ -5,6 +5,7 @@
 package org.opensearch.neuralsearch.query.visitor;
 
 import java.util.Map;
+import java.util.Objects;
 
 import org.apache.lucene.search.BooleanClause;
 import org.opensearch.index.query.QueryBuilder;
@@ -22,6 +23,13 @@ public class NeuralSearchQueryVisitor implements QueryBuilderVisitor {
     private final String modelId;
     private final Map<String, Object> neuralFieldMap;
 
+    private Boolean isFieldDefaultModelIdApplied(String fieldName) {
+        if (Objects.nonNull(neuralFieldMap) && Objects.nonNull(fieldName) && Objects.nonNull(neuralFieldMap.get(fieldName))) {
+            return true;
+        }
+        return false;
+    }
+
     /**
      * Accept method accepts every query builder from the search request,
      * and processes it if the required conditions in accept method are satisfied.
@@ -31,12 +39,10 @@ public class NeuralSearchQueryVisitor implements QueryBuilderVisitor {
         if (queryBuilder instanceof ModelInferenceQueryBuilder) {
             ModelInferenceQueryBuilder modelInferenceQueryBuilder = (ModelInferenceQueryBuilder) queryBuilder;
             if (modelInferenceQueryBuilder.modelId() == null) {
-                if (neuralFieldMap != null
-                    && modelInferenceQueryBuilder.fieldName() != null
-                    && neuralFieldMap.get(modelInferenceQueryBuilder.fieldName()) != null) {
+                if (isFieldDefaultModelIdApplied(modelInferenceQueryBuilder.fieldName())) {
                     String fieldDefaultModelId = (String) neuralFieldMap.get(modelInferenceQueryBuilder.fieldName());
                     modelInferenceQueryBuilder.modelId(fieldDefaultModelId);
-                } else if (modelId != null) {
+                } else if (Objects.nonNull(modelId)) {
                     modelInferenceQueryBuilder.modelId(modelId);
                 } else {
                     throw new IllegalArgumentException(